[Fix] 커밋 충돌 해결 및 통신 구현
